update: moved all user crud or services to its own module
